requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
biopython>=1.81
scikit-allel>=1.3.0
polars>=0.19.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
cyvcf2
myvariant
pypgx
SQLAlchemy
pydantic
ollama
langchain
langchain-community
chromadb